自动提交 git 代码

+++ hander/brand.go
@@ -2,6 +2,7 @@ package hander
 
 import (
 	"context"
+	"fmt"
 
 	pb "github.com/gomsa/goods/proto/brand"
 	"github.com/gomsa/goods/service"
@@ -14,6 +15,10 @@ type Brand struct {
 
 // Exist 获取所有品牌
 func (srv *Brand) Exist(ctx context.Context, req *pb.Request, res *pb.Response) (err error) {
+	// 防止空指针报错
+	if req.Brand == nil {
+		req.Brand = &pb.Brand{}
+	}
 	valid, err := srv.Repo.Exist(req.Brand)
 	if err != nil {
 		return err
@@ -24,6 +29,10 @@ func (srv *Brand) Exist(ctx context.Context, req *pb.Request, res *pb.Response) 
 
 // All 获取所有品牌
 func (srv *Brand) All(ctx context.Context, req *pb.Request, res *pb.Response) (err error) {
+	// 防止空指针报错
+	if req.Brand == nil {
+		req.Brand = &pb.Brand{}
+	}
 	brands, err := srv.Repo.All(req.Brand)
 	if err != nil {
 		return err
@@ -34,6 +43,13 @@ func (srv *Brand) All(ctx context.Context, req *pb.Request, res *pb.Response) (e
 
 // List 获取所有品牌
 func (srv *Brand) List(ctx context.Context, req *pb.Request, res *pb.Response) (err error) {
+	// 防止空指针报错
+	if req.ListQuery == nil {
+		req.ListQuery = &pb.ListQuery{}
+	}
+	if req.Brand == nil {
+		req.Brand = &pb.Brand{}
+	}
 	brands, err := srv.Repo.List(req.ListQuery, req.Brand)
 	if err != nil {
 		return err
@@ -44,6 +60,10 @@ func (srv *Brand) List(ctx context.Context, req *pb.Request, res *pb.Response) (
 
 // Get 获取品牌
 func (srv *Brand) Get(ctx context.Context, req *pb.Request, res *pb.Response) (err error) {
+	// 防止空指针报错
+	if req.Brand == nil {
+		return fmt.Errorf("请求参数错误")
+	}
 	brand, err := srv.Repo.Get(req.Brand)
 	if err != nil {
 		return err
@@ -54,6 +74,10 @@ func (srv *Brand) Get(ctx context.Context, req *pb.Request, res *pb.Response) (e
 
 // Create 创建品牌
 func (srv *Brand) Create(ctx context.Context, req *pb.Request, res *pb.Response) (err error) {
+	// 防止空指针报错
+	if req.Brand == nil {
+		return fmt.Errorf("请求参数错误")
+	}
 	brand, err := srv.Repo.Create(req.Brand)
 	if err != nil {
 		res.Valid = false
@@ -66,6 +90,10 @@ func (srv *Brand) Create(ctx context.Context, req *pb.Request, res *pb.Response)
 
 // Update 更新品牌
 func (srv *Brand) Update(ctx context.Context, req *pb.Request, res *pb.Response) (err error) {
+	// 防止空指针报错
+	if req.Brand == nil {
+		return fmt.Errorf("请求参数错误")
+	}
 	valid, err := srv.Repo.Update(req.Brand)
 	if err != nil {
 		res.Valid = false
@@ -77,6 +105,10 @@ func (srv *Brand) Update(ctx context.Context, req *pb.Request, res *pb.Response)
 
 // Delete 删除品牌品牌
 func (srv *Brand) Delete(ctx context.Context, req *pb.Request, res *pb.Response) (err error) {
+	// 防止空指针报错
+	if req.Brand == nil {
+		return fmt.Errorf("请求参数错误")
+	}
 	valid, err := srv.Repo.Delete(req.Brand)
 	if err != nil {
 		res.Valid = false
