+++ hander/firm.go
@@ -2,6 +2,7 @@ package hander
 
 import (
 	"context"
+	"fmt"
 
 	pb "github.com/gomsa/goods/proto/firm"
 	"github.com/gomsa/goods/service"
@@ -14,6 +15,10 @@ type Firm struct {
 
 // Exist 获取所有商品公司
 func (srv *Firm) Exist(ctx context.Context, req *pb.Request, res *pb.Response) (err error) {
+	// 防止空指针报错
+	if req.Firm == nil {
+		return fmt.Errorf("请求参数错误")
+	}
 	valid, err := srv.Repo.Exist(req.Firm)
 	if err != nil {
 		return err
@@ -24,6 +29,10 @@ func (srv *Firm) Exist(ctx context.Context, req *pb.Request, res *pb.Response) (
 
 // All 获取所有商品公司
 func (srv *Firm) All(ctx context.Context, req *pb.Request, res *pb.Response) (err error) {
+	// 防止空指针报错
+	if req.Firm == nil {
+		req.Firm = &pb.Firm{}
+	}
 	firms, err := srv.Repo.All(req.Firm)
 	if err != nil {
 		return err
@@ -34,6 +43,13 @@ func (srv *Firm) All(ctx context.Context, req *pb.Request, res *pb.Response) (er
 
 // List 获取所有商品公司
 func (srv *Firm) List(ctx context.Context, req *pb.Request, res *pb.Response) (err error) {
+	// 防止空指针报错
+	if req.Firm == nil {
+		req.Firm = &pb.Firm{}
+	}
+	if req.ListQuery == nil {
+		req.ListQuery = &pb.ListQuery{}
+	}
 	firms, err := srv.Repo.List(req.ListQuery, req.Firm)
 	if err != nil {
 		return err
@@ -44,6 +60,10 @@ func (srv *Firm) List(ctx context.Context, req *pb.Request, res *pb.Response) (e
 
 // Get 获取商品公司
 func (srv *Firm) Get(ctx context.Context, req *pb.Request, res *pb.Response) (err error) {
+	// 防止空指针报错
+	if req.Firm == nil {
+		return fmt.Errorf("请求参数错误")
+	}
 	firm, err := srv.Repo.Get(req.Firm)
 	if err != nil {
 		return err
@@ -54,6 +74,10 @@ func (srv *Firm) Get(ctx context.Context, req *pb.Request, res *pb.Response) (er
 
 // Create 创建商品公司
 func (srv *Firm) Create(ctx context.Context, req *pb.Request, res *pb.Response) (err error) {
+	// 防止空指针报错
+	if req.Firm == nil {
+		return fmt.Errorf("请求参数错误")
+	}
 	firm, err := srv.Repo.Create(req.Firm)
 	if err != nil {
 		res.Valid = false
@@ -66,6 +90,10 @@ func (srv *Firm) Create(ctx context.Context, req *pb.Request, res *pb.Response) 
 
 // Update 更新商品公司
 func (srv *Firm) Update(ctx context.Context, req *pb.Request, res *pb.Response) (err error) {
+	// 防止空指针报错
+	if req.Firm == nil {
+		return fmt.Errorf("请求参数错误")
+	}
 	valid, err := srv.Repo.Update(req.Firm)
 	if err != nil {
 		res.Valid = false
@@ -77,6 +105,10 @@ func (srv *Firm) Update(ctx context.Context, req *pb.Request, res *pb.Response) 
 
 // Delete 删除商品公司商品公司
 func (srv *Firm) Delete(ctx context.Context, req *pb.Request, res *pb.Response) (err error) {
+	// 防止空指针报错
+	if req.Firm == nil {
+		return fmt.Errorf("请求参数错误")
+	}
 	valid, err := srv.Repo.Delete(req.Firm)
 	if err != nil {
 		res.Valid = false
