+++ hander/taxcode.go
@@ -2,6 +2,7 @@ package hander
 
 import (
 	"context"
+	"fmt"
 
 	pb "github.com/gomsa/goods/proto/taxcode"
 	"github.com/gomsa/goods/service"
@@ -14,6 +15,10 @@ type Taxcode struct {
 
 // All 获取所有国家税收分类编码
 func (srv *Taxcode) All(ctx context.Context, req *pb.Request, res *pb.Response) (err error) {
+	// 防止空指针报错
+	if req.Taxcode == nil {
+		req.Taxcode = &pb.Taxcode{}
+	}
 	taxcodes, err := srv.Repo.All(req.Taxcode)
 	if err != nil {
 		return err
@@ -24,6 +29,10 @@ func (srv *Taxcode) All(ctx context.Context, req *pb.Request, res *pb.Response) 
 
 // List 获取所有国家税收分类编码
 func (srv *Taxcode) List(ctx context.Context, req *pb.Request, res *pb.Response) (err error) {
+	// 防止空指针报错
+	if req.Taxcode == nil {
+		req.Taxcode = &pb.Taxcode{}
+	}
 	taxcodes, err := srv.Repo.List(req.Taxcode)
 	if err != nil {
 		return err
@@ -34,6 +43,10 @@ func (srv *Taxcode) List(ctx context.Context, req *pb.Request, res *pb.Response)
 
 // Get 获取国家税收分类编码
 func (srv *Taxcode) Get(ctx context.Context, req *pb.Request, res *pb.Response) (err error) {
+	// 防止空指针报错
+	if req.Taxcode == nil {
+		return fmt.Errorf("请求参数错误")
+	}
 	taxcode, err := srv.Repo.Get(req.Taxcode)
 	if err != nil {
 		return err
@@ -44,6 +57,10 @@ func (srv *Taxcode) Get(ctx context.Context, req *pb.Request, res *pb.Response) 
 
 // Create 创建国家税收分类编码
 func (srv *Taxcode) Create(ctx context.Context, req *pb.Request, res *pb.Response) (err error) {
+	// 防止空指针报错
+	if req.Taxcode == nil {
+		return fmt.Errorf("请求参数错误")
+	}
 	taxcode, err := srv.Repo.Create(req.Taxcode)
 	if err != nil {
 		res.Valid = false
@@ -56,6 +73,10 @@ func (srv *Taxcode) Create(ctx context.Context, req *pb.Request, res *pb.Respons
 
 // Update 更新国家税收分类编码
 func (srv *Taxcode) Update(ctx context.Context, req *pb.Request, res *pb.Response) (err error) {
+	// 防止空指针报错
+	if req.Taxcode == nil {
+		return fmt.Errorf("请求参数错误")
+	}
 	valid, err := srv.Repo.Update(req.Taxcode)
 	if err != nil {
 		res.Valid = false
@@ -67,6 +88,10 @@ func (srv *Taxcode) Update(ctx context.Context, req *pb.Request, res *pb.Respons
 
 // Delete 删除国家税收分类编码国家税收分类编码
 func (srv *Taxcode) Delete(ctx context.Context, req *pb.Request, res *pb.Response) (err error) {
+	// 防止空指针报错
+	if req.Taxcode == nil {
+		return fmt.Errorf("请求参数错误")
+	}
 	valid, err := srv.Repo.Delete(req.Taxcode)
 	if err != nil {
 		res.Valid = false
